Moved actions worker
